# Speedup parsing 2.6x