--- a/lexer/lexer.go
+++ b/lexer/lexer.go
@@ -21,6 +21,10 @@ const (
 	closeMustache               = "}}"
 	closeStripMustache          = "~}}"
 	closeUnescapedStripMustache = "}~}}"
+	openComment                 = "{{!"
+	openCommentStrip            = "{{~!"
+	openCommentDash             = "{{!--"
+	openCommentStripDash        = "{{~!--"
 )
 
 const eof = -1
@@ -74,11 +78,9 @@ var (
 	rOpen            = regexp.MustCompile(`^\{\{~?&?`)
 	rClose           = regexp.MustCompile(`^~?\}\}`)
 	rOpenBlockParams = regexp.MustCompile(`^as\s+\|`)
-	// {{!--  ... --}}
-	rOpenCommentDash  = regexp.MustCompile(`^\{\{~?!--\s*`)
+	// --}} or --~}}
 	rCloseCommentDash = regexp.MustCompile(`^\s*--~?\}\}`)
-	// {{! ... }}
-	rOpenComment  = regexp.MustCompile(`^\{\{~?!\s*`)
+	// }} or ~}}
 	rCloseComment = regexp.MustCompile(`^\s*~?\}\}`)
 )
 
@@ -278,12 +280,12 @@ func lexContent(l *Lexer) lexFunc {
 	} else if l.isString(escapedOpenMustache) {
 		// \{{
 		next = lexEscapedOpenMustache
-	} else if str := l.findRegexp(rOpenCommentDash); str != "" {
+	} else if l.isString(openCommentDash) || l.isString(openCommentStripDash) {
 		// {{!--
 		l.closeComment = rCloseCommentDash
 
 		next = lexComment
-	} else if str := l.findRegexp(rOpenComment); str != "" {
+	} else if l.isString(openComment) || l.isString(openCommentStrip) {
 		// {{!
 		l.closeComment = rCloseComment