Hello @imantung, I am using your library in https://github.com/xvello/letsblockit and noticed that, although the rendering performance is great, the parsing speed can be improved.. A quick profile, using [the 11 handlebar files in my project](https://github.com/xvello/letsblockit/tree/main/data/pages) showed some quick-wins that I want to contribute back.

This first PR improves `lexContent` (the hottest function in my initial profiling), resulting in a **2.6x speed increase** on my (admitedly small) test corpus. I'd be happy to re-run the parsing benchmark with another corpus if you can provide one.

## Profiles 
Here are PDF renderings of [initial profiling](https://github.com/imantung/mario/files/7597897/cpu_orig.pdf) and [profiling of this branch](https://github.com/imantung/mario/files/7597898/cpu_patched.pdf). There are more branches to optimize, but I wanted to start with an easy PR.

## Benchmarks

#### Before

```
± go test ./bench/ -bench=. -benchtime=30s -benchmem -cpuprofile cpu.prof
goos: linux
goarch: amd64
pkg: github.com/imantung/mario/bench
cpu: AMD Ryzen 5 PRO 3500U w/ Radeon Vega Mobile Gfx
BenchmarkPageLoad-8         1220          31088188 ns/op          434625 B/op       6324 allocs/op
BenchmarkPageLoad-8         1213          31095209 ns/op          434913 B/op       6324 allocs/op
BenchmarkPageLoad-8         1222          30983005 ns/op          433669 B/op       6323 allocs/op
```
#### After replacing the regexp with `isString`

```
BenchmarkPageLoad-8         2740          13910230 ns/op          431941 B/op       6323 allocs/op
BenchmarkPageLoad-8         2754          13988925 ns/op          431047 B/op       6322 allocs/op
```
#### After adding peek
```
BenchmarkPageLoad-8         3288          11679414 ns/op          428415 B/op       6322 allocs/op
BenchmarkPageLoad-8         3526          11795542 ns/op          430295 B/op       6323 allocs/op
BenchmarkPageLoad-8         4036          11716324 ns/op          428726 B/op       6322 allocs/op
```